frontend: Load config file before app initialization
This is necessary for Matomo, because I want to outsource the matomo configuration to the docker-compose rather than having constant values in the `app.module.ts`.
The problem of our previous way of loading the config file was that it was loaded during app initialization and Matomo comes as a module which are loaded at the same time.
So it was not possible to provide the configuration for the Matomo module, because the config was not available yet.